refactor: (#14) package structure
