+++ dms-application/src/main/kotlin/team/aliens/dms/domain/user/spi/UserJwtPort.kt
@@ -1,6 +1,6 @@
 package team.aliens.dms.domain.user.spi
 
-import team.aliens.dms.domain.auth.dto.response.TokenResponse
+import team.aliens.dms.domain.auth.dto.TokenResponse
 import team.aliens.dms.domain.auth.model.Authority
 import java.util.*
